# fix: order - AOP 코드 수정.
AOP 수정 및
memberId를 myUserDetail에서
가져오기

+++ src/main/java/com/example/minzok/order/service/OrderService.java
@@ -86,7 +86,7 @@ public class OrderService {
         return OrderResponseDto.builder()
                 .orderId(order.getId())
                 .storeId(store.getId())
-                .totalPrice(Long.valueOf(Math.toIntExact(order.getTotalPrice())))
+                .totalPrice(Long.valueOf((order.getTotalPrice())))
                 .orderStatus(order.getOrderStatus().name())
                 .orderTime(order.getOrderTime())
                 .menus(List.of(menuDto))
